:art: typescriptify recognize_stream and refactor TTS adapter
